chore: Reorganized to allow multiple flows
