+++ src/lib/2.app/driver-ports/dynamic-init/for-processing-remote-entry.port.ts
@@ -1,0 +1,3 @@
+import type { RemoteEntry } from 'lib/1.domain/remote-entry/remote-entry.contract';
+
+export type ForProcessingRemoteEntry = (remoteEntries: RemoteEntry) => Promise<void>;
