+++ src/lib/2.app/driver-ports/init/for-determining-shared-externals.port.ts
