+++ src/lib/2.app/flows/init/generate-import-map.shared-externals.spec.ts
@@ -1,11 +1,11 @@
-import { ForGeneratingImportMap } from '../driver-ports/for-generating-import-map';
-import { DrivingContract } from '../driving-ports/driving.contract';
+import { ForGeneratingImportMap } from '../../driver-ports/init/for-generating-import-map';
+import { DrivingContract } from '../../driving-ports/driving.contract';
 import { createGenerateImportMap } from './generate-import-map';
 import { mockSharedExternalsRepository } from 'lib/6.mocks/adapters/shared-externals.repository.mock';
 import { mockScopedExternalsRepository } from 'lib/6.mocks/adapters/scoped-externals.repository.mock';
 import { mockRemoteInfoRepository } from 'lib/6.mocks/adapters/remote-info.repository.mock';
-import { LoggingConfig } from '../config/log.contract';
-import { ModeConfig } from '../config/mode.contract';
+import { LoggingConfig } from '../../config/log.contract';
+import { ModeConfig } from '../../config/mode.contract';
 import { NFError } from 'lib/native-federation.error';
 
 describe('createGenerateImportMap (shared-externals)', () => {
