+++ src/lib/2.app/flows/init/generate-import-map.ts
@@ -1,9 +1,9 @@
-import type { ForGeneratingImportMap } from '../driver-ports/for-generating-import-map';
-import type { DrivingContract } from '../driving-ports/driving.contract';
+import type { ForGeneratingImportMap } from '../../driver-ports/init/for-generating-import-map';
+import type { DrivingContract } from '../../driving-ports/driving.contract';
 import type { ImportMap, Imports } from 'lib/1.domain/import-map/import-map.contract';
 import type { ExternalsScope, RemoteInfo, SharedExternal, SharedVersion } from 'lib/1.domain';
-import type { LoggingConfig } from '../config/log.contract';
-import type { ModeConfig } from '../config/mode.contract';
+import type { LoggingConfig } from '../../config/log.contract';
+import type { ModeConfig } from '../../config/mode.contract';
 import * as _path from 'lib/utils/path';
 import { NFError } from 'lib/native-federation.error';
