+++ src/lib/2.app/flows/init/get-remote-entries.spec.ts
@@ -1,6 +1,6 @@
 import { createGetRemoteEntries } from './get-remote-entries';
-import { DrivingContract } from '../driving-ports/driving.contract';
-import { ForGettingRemoteEntries } from '../driver-ports/for-getting-remote-entries.port';
+import { DrivingContract } from '../../driving-ports/driving.contract';
+import { ForGettingRemoteEntries } from '../../driver-ports/init/for-getting-remote-entries.port';
 import { mockManifestProvider } from 'lib/6.mocks/adapters/manifest-provider.mock';
 import { mockRemoteEntryProvider } from 'lib/6.mocks/adapters/remote-entry-provider.mock';
 import { mockRemoteInfoRepository } from 'lib/6.mocks/adapters/remote-info.repository.mock';
@@ -14,9 +14,9 @@ import {
   MOCK_REMOTE_ENTRY_SCOPE_II_URL,
 } from 'lib/6.mocks/domain/remote-entry/remote-entry.mock';
 import { NFError } from 'lib/native-federation.error';
-import { HostConfig } from '../config/host.contract';
-import { LoggingConfig } from '../config/log.contract';
-import { ModeConfig } from '../config/mode.contract';
+import { HostConfig } from '../../config/host.contract';
+import { LoggingConfig } from '../../config/log.contract';
+import { ModeConfig } from '../../config/mode.contract';
 
 describe('createGetRemoteEntries', () => {
   let getRemoteEntries: ForGettingRemoteEntries;
