+++ src/lib/2.app/flows/init/index.ts
@@ -1,0 +1,1 @@
+export * from '../../config';
