+++ src/lib/2.app/flows/init/process-remote-entries.spec.ts
@@ -1,7 +1,7 @@
 import { createProcessRemoteEntries } from './process-remote-entries';
-import { ForProcessingRemoteEntries } from '../driver-ports/for-processing-remote-entries.port';
-import { DrivingContract } from '../driving-ports/driving.contract';
-import { LoggingConfig } from '../config/log.contract';
+import { ForProcessingRemoteEntries } from '../../driver-ports/init/for-processing-remote-entries.port';
+import { DrivingContract } from '../../driving-ports/driving.contract';
+import { LoggingConfig } from '../../config/log.contract';
 import { mockRemoteInfoRepository } from 'lib/6.mocks/adapters/remote-info.repository.mock';
 import { mockSharedExternalsRepository } from 'lib/6.mocks/adapters/shared-externals.repository.mock';
 import { mockScopedExternalsRepository } from 'lib/6.mocks/adapters/scoped-externals.repository.mock';
