+++ src/lib/2.app/flows/init/process-remote-entries.ts
@@ -1,7 +1,7 @@
-import type { ForProcessingRemoteEntries } from '../driver-ports/for-processing-remote-entries.port';
+import type { ForProcessingRemoteEntries } from '../../driver-ports/init/for-processing-remote-entries.port';
 import type { RemoteEntry, RemoteInfo, SharedInfo, SharedVersion, Version } from 'lib/1.domain';
-import type { DrivingContract } from '../driving-ports/driving.contract';
-import type { LoggingConfig } from '../config/log.contract';
+import type { DrivingContract } from '../../driving-ports/driving.contract';
+import type { LoggingConfig } from '../../config/log.contract';
 import * as _path from 'lib/utils/path';
 
 export function createProcessRemoteEntries(
