--- src/lib/2.app/init-federation/index.ts
@@ -1,1 +1,0 @@
-export * from '../config';
