+++ src/lib/create-nf-app.ts
@@ -1,8 +1,8 @@
 import type { ConfigContract, NFOptions } from './2.app/config/config.contract';
-import type { DriversContract } from './2.app/driver-ports/drivers.contract';
+import type { DriversContract } from './2.app/driver-ports/init/drivers.contract';
 import type { DrivingContract } from './2.app/driving-ports/driving.contract';
 import { createConfigHandlers } from './5.di/config.factory';
-import { createDrivers } from './5.di/drivers.factory';
+import { createInitDrivers } from './5.di/init-drivers.factory';
 import { createDriving } from './5.di/driving.factory';
 
 export type NF_APP = {
@@ -14,7 +14,7 @@ export type NF_APP = {
 export const CREATE_NF_APP = (options: NFOptions): NF_APP => {
   const config = createConfigHandlers(options);
   const adapters = createDriving(config);
-  const app = createDrivers(config, adapters);
+  const app = createInitDrivers(config, adapters);
 
   return {
     app,
