+++ src/lib/index.ts
@@ -2,4 +2,4 @@ export { initFederation } from './init-federation';
 export { NFError } from './native-federation.error';
 
 export * from './1.domain';
-export * from './2.app/init-federation';
+export * from './init-federation';
