+++ src/lib/sdk.index.ts
@@ -1,4 +1,4 @@
 export * from './create-nf-app';
 
-export * from './2.app/driver-ports';
+export * from './2.app/driver-ports/init';
 export * from './2.app/driving-ports';
